[onchain-discovery] move `OptionFuture` into common/
