+++ network/onchain-discovery/Cargo.toml
@@ -14,6 +14,7 @@ anyhow = "1.0"
 bytes = "0.5.4"
 futures = "0.3.4"
 once_cell = "1.3.1"
+option-future = { path = "../../common/option-future", version = "0.1.0" }
 rand = "0.7.3"
 serde = { version = "1.0.106", features = ["derive"] }
 tokio = { version = "0.2.13", features = ["full"] }
