+++ network/onchain-discovery/src/lib.rs
@@ -51,13 +51,8 @@ use crate::types::{
     QueryDiscoverySetRequest, QueryDiscoverySetResponse, QueryDiscoverySetResponseWithEvent,
 };
 use anyhow::{Context as AnyhowContext, Result};
-use futures::{
-    future::{FusedFuture, Future},
-    ready,
-    task::Poll,
-};
 use libra_types::get_with_proof::{UpdateToLatestLedgerRequest, UpdateToLatestLedgerResponse};
-use std::{convert::TryFrom, pin::Pin, sync::Arc, task::Context};
+use std::{convert::TryFrom, sync::Arc};
 use storage_client::StorageRead;
 
 #[cfg(test)]
@@ -106,63 +101,4 @@ async fn storage_query_discovery_set(
     })?;
 
     Ok((req_msg, res_msg))
-}
-
-/// Effectively a size=1 future queue, or a slot with at-most-one future inside.
-///
-/// For now, we require `F: Unpin` which mandates boxing of `async fn` futures
-/// but simplifies the `OptionFuture` implementation.
-// TODO(philiphayes): extract this into libra/common
-pub struct OptionFuture<F> {
-    inner: Option<F>,
-}
-
-impl<F> OptionFuture<F>
-where
-    F: Future + Unpin,
-{
-    pub fn new(inner: Option<F>) -> Self {
-        Self { inner }
-    }
-
-    pub fn or_insert_with<G>(&mut self, fun: G)
-    where
-        G: FnOnce() -> Option<F>,
-    {
-        if self.inner.is_none() {
-            self.inner = fun();
-        }
-    }
-}
-
-impl<F> Unpin for OptionFuture<F> where F: Unpin {}
-
-impl<F> Future for OptionFuture<F>
-where
-    F: Future + Unpin,
-{
-    type Output = <F as Future>::Output;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<Self::Output> {
-        // try to poll whatever future is in the inner slot, if there is one.
-        let out = match Pin::new(&mut self.as_mut().get_mut().inner).as_pin_mut() {
-            Some(f) => ready!(f.poll(cx)),
-            None => return Poll::Pending,
-        };
-
-        // the inner future is complete so we can drop it now and just return the
-        // results.
-        self.set(OptionFuture { inner: None });
-
-        Poll::Ready(out)
-    }
-}
-
-impl<F> FusedFuture for OptionFuture<F>
-where
-    F: Future + Unpin,
-{
-    fn is_terminated(&self) -> bool {
-        self.inner.is_none()
-    }
 }
